requirements.txt
# DSPy framework
dspy-ai>=2.4.0

# Dataset loading
datasets>=2.14.0

# Ollama integration
ollama>=0.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0

# Optional: for better visualization
matplotlib>=3.7.0
seaborn>=0.12.0
